Completed Quiz feature with caching and repository layer.
