+++ app/src/main/java/com/example/mycheesecakes/data/cache/model/CachedQuizResultWithQuestions.kt
@@ -1,0 +1,9 @@
+package com.example.mycheesecakes.data.cache.model
+
+/**
+ * Used to view correct/incorrect questions in Quiz Results page.
+ */
+data class CachedQuizResultWithQuestions(
+    val quizResult: CachedQuizResult,
+    val questions: List<CachedQuestion>
+)
