+++ app/src/main/java/com/example/mycheesecakes/data/network/api/model/mappers/ApiQuizResultMapper.kt
@@ -1,7 +1,6 @@
 package com.example.mycheesecakes.data.network.api.model.mappers
 
 import com.example.mycheesecakes.data.network.api.model.ApiQuizResult
-import com.example.mycheesecakes.domain.model.Quiz
 import com.example.mycheesecakes.domain.model.quiz.QuizResult
 import com.example.mycheesecakes.domain.model.quiz.Score
 
@@ -11,7 +10,8 @@ class ApiQuizResultMapper: ApiMapper<ApiQuizResult,QuizResult> {
             score = Score(
                 correct = apiEntity?.fields?.correct ?: 0,
                 incorrect = apiEntity?.fields?.incorrect ?: 0),
-            date = apiEntity?.fields?.date ?: 0,
-            id = apiEntity?.fields?.id ?: 0)
+            date = apiEntity.fields?.date ?: 0,
+            id = apiEntity.fields?.id ?: 0,
+        quizId = apiEntity.fields?.quizId ?: -1)
     }
 }
