+++ app/src/main/java/com/example/mycheesecakes/domain/model/quiz/QuizResult.kt
@@ -1,10 +1,9 @@
 package com.example.mycheesecakes.domain.model.quiz
 
-import java.util.*
-
 data class QuizResult(
     val score: Score,
     val date: Long,
-    val id: Int? = null
+    val id: Int = 0,
+    val quizId: Int
 ) {
 }
