+++ app/src/main/java/com/example/mycheesecakes/ui/flashcards/flashcardssetup/FlashcardsSetupFragment.kt
@@ -1,14 +1,11 @@
 package com.example.mycheesecakes.ui.flashcards.flashcardssetup
 
-import android.graphics.Canvas
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
-import androidx.recyclerview.widget.RecyclerView
 import com.example.mycheesecakes.databinding.FragmentFlashcardsSetupBinding
 import com.example.mycheesecakes.domain.model.menuitems.menuItems
 
@@ -16,8 +13,7 @@ import com.example.mycheesecakes.domain.model.menuitems.menuItems
 class FlashcardsSetupFragment : Fragment() {
 
     private lateinit var binding: FragmentFlashcardsSetupBinding
-    private val cheesecakes = menuItems
-    private val adapter = FlashcardsSetupAdapter(cheesecakes)
+    private val adapter = FlashcardsSetupAdapter(menuItems)
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
