+++ app/src/main/java/com/example/mycheesecakes/ui/quiz/QuizFragment.kt
@@ -1,5 +1,6 @@
 package com.example.mycheesecakes.ui.quiz
 
+import android.graphics.Color
 import android.os.Build
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -9,9 +10,15 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
 import androidx.annotation.RequiresApi
+import androidx.navigation.fragment.findNavController
 import com.bumptech.glide.Glide
 import com.example.mycheesecakes.R
 import com.example.mycheesecakes.databinding.FragmentQuizBinding
+import com.google.android.material.snackbar.Snackbar
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.collect
+import kotlinx.coroutines.launch
 
 class QuizFragment : Fragment(), View.OnClickListener {
 
@@ -53,6 +60,22 @@ class QuizFragment : Fragment(), View.OnClickListener {
         viewModel.timeRemaining.observe(viewLifecycleOwner) { time ->
             binding.countdownTimerTextView.text = time.toString()
         }
+
+
+        viewModel.answerResult.observe(viewLifecycleOwner) { result ->
+            showAnswerSnackbar(result)
+        }
+
+        viewModel.quizScore.observe(viewLifecycleOwner) { score ->
+            val action = QuizFragmentDirections.actionQuizFragmentToQuizResultsSimpleFragment(score.correct,(score.correct + score.incorrect))
+            findNavController().navigate(action)
+        }
+
+        viewModel.errorRetrievingData.observe(viewLifecycleOwner) { error ->
+            if (error) {
+                Toast.makeText(context,"Unable to create quiz. Check your internet connection and try again.", Toast.LENGTH_LONG).show()
+            }
+        }
     }
 
     override fun onStop() {
@@ -88,11 +111,12 @@ class QuizFragment : Fragment(), View.OnClickListener {
     }
 
 
-    private fun showAnswerResultToast(correct: Boolean) {
-        if (correct) {
-            Toast.makeText(context,"Correct!",Toast.LENGTH_SHORT).show()
-        } else {
-            //Toast.makeText(context,"Wrong:\n${viewModel.correctAnswer}", Toast.LENGTH_SHORT).show()
+    private fun showAnswerSnackbar(result: String) {
+        context?.let { context ->
+            Snackbar.make(context, binding.root, result, Snackbar.LENGTH_SHORT)
+                .setBackgroundTint(Color.parseColor("#FEFBF7E6"))
+                .setTextColor(Color.parseColor("#824D3B"))
+                .show()
         }
     }
 
@@ -109,40 +133,6 @@ class QuizFragment : Fragment(), View.OnClickListener {
         binding.answer4Button.setOnClickListener(this)
 
     }
-    /*
-
-    private fun observeLiveData() {
-        viewModel.cheesecakeLiveData.observe(viewLifecycleOwner, { cheesecake ->
-            binding.cheesecakeNameTextview.text = cheesecake.name
-            Glide
-                    .with(this)
-                    .load(cheesecake.imageURL)
-                    .centerCrop()
-                    .placeholder(R.drawable.fresh_strawberry_cheesecake)
-                    .into(binding.cheesecakeImageview)
-        })
-
-        viewModel.questionLiveData.observe(viewLifecycleOwner, { question ->
-            binding.cheesecakeQuestionTextview.text = question
-        })
-
-        viewModel.answers.observe(viewLifecycleOwner, { answers ->
-            binding.answer1Button.text = answers[0]
-            binding.answer2Button.text = answers[1]
-            binding.answer3Button.text = answers[2]
-            binding.answer4Button.text = answers[3]
-        })
-
-        viewModel.correctAnswerLiveData.observe(viewLifecycleOwner, { result ->
-            result?.let {
-                showAnswerResultToast(result)
-            }
-        })
-
-        )
-    }
-
-     */
 
     override fun onPause() {
         super.onPause()
