+++ app/src/main/java/com/example/mycheesecakes/ui/quiz/QuizViewModel.kt
@@ -1,32 +1,32 @@
 package com.example.mycheesecakes.ui.quiz
 
-import android.os.Build
 import android.util.Log
-import androidx.annotation.RequiresApi
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.mycheesecakes.data.network.RetrofitInstance
-import com.example.mycheesecakes.data.network.api.model.mappers.*
-import com.example.mycheesecakes.domain.model.Cheesecake
+import com.example.mycheesecakes.app.App.Companion.menuItemRepository
+import com.example.mycheesecakes.app.App.Companion.quizRepository
+import com.example.mycheesecakes.data.repositories.MenuItemRepository
+import com.example.mycheesecakes.data.repositories.QuizRepository
 import com.example.mycheesecakes.domain.model.Quiz
-import com.example.mycheesecakes.domain.model.menuitems.Dessert
-import com.example.mycheesecakes.domain.model.menuitems.Drink
 import com.example.mycheesecakes.domain.model.menuitems.MenuItem
 import com.example.mycheesecakes.domain.model.quiz.Question
 import com.example.mycheesecakes.domain.model.quiz.QuizProvider
-import com.example.mycheesecakes.domain.model.quiz.QuizResult
+import com.example.mycheesecakes.domain.model.quiz.Score
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.collect
 import kotlinx.coroutines.launch
-import retrofit2.HttpException
-import java.io.IOException
 import java.lang.IllegalArgumentException
 
 const val TAG = "QuizViewModel"
 
-class QuizViewModel(menuItemType: Int) : ViewModel() {
+class QuizViewModel(
+    menuItemType: Int,
+    quizRepository: QuizRepository,
+    menuItemRepository: MenuItemRepository
+) : ViewModel() {
 
     private lateinit var quiz: Quiz
     private var question: Question? = null
@@ -40,27 +40,55 @@ class QuizViewModel(menuItemType: Int) : ViewModel() {
     val timeRemaining: LiveData<Int>
     get() = _timeRemaining
 
+    private val _answerResult = MutableLiveData<String>()
+    val answerResult: LiveData<String>
+    get() = _answerResult
+
+    private val _quizScore = MutableLiveData<Score>()
+    val quizScore: LiveData<Score>
+    get() = _quizScore
+
+    private val _errorRetrievingData = MutableLiveData<Boolean>()
+    val errorRetrievingData: LiveData<Boolean>
+    get() = _errorRetrievingData
+
     private var quizIsStarted = false
 
     init {
-        val menuItems = getMenuItems(menuItemType)
-        Log.i(TAG,menuItems.toString())
-
+        viewModelScope.launch {
+            val cachedQuiz = quizRepository.getUnfinishedQuiz()
+            if (cachedQuiz != null) {
+                quiz = cachedQuiz
+                beginQuiz()
+            } else {
+                val menuItems = getMenuItems(menuItemType)
+                if (menuItems != null) {
+                    setupNewQuiz(menuItems)
+                } else {
+                    errorRetrievingItems()
+                }
+            }
+        }
     }
 
-    private fun getMenuItems(menuItemType: Int): List<MenuItem> {
+    private suspend fun getMenuItems(menuItemType: Int): List<MenuItem>? {
         return when (menuItemType) {
-            MenuItem.TYPE_CHEESECAKE -> getCheesecakes()
-            MenuItem.TYPE_DESSERT -> getDesserts()
-            MenuItem.TYPE_DRINK -> getDrinks()
+            MenuItem.TYPE_CHEESECAKE -> menuItemRepository.getAllCheesecakes()
+            MenuItem.TYPE_DESSERT -> menuItemRepository.getAllDesserts()
+            MenuItem.TYPE_DRINK -> menuItemRepository.getAllDrinks()
             else -> throw IllegalArgumentException("Invalid menuItemType")
         }
     }
 
-    private fun setupQuiz(menuItems: List<MenuItem>) {
-        quiz = QuizProvider().createQuiz(menuItems,QuizProvider.QUIZ_SIZE_LARGE)
+    private fun beginQuiz() {
         quizIsStarted = true
         nextQuestion()
+    }
+
+    private fun setupNewQuiz(menuItems: List<MenuItem>, quizSize: Int = QuizProvider.QUIZ_SIZE_MEDIUM) {
+        quiz = QuizProvider().createQuiz(menuItems,quizSize)
+        //viewModelScope.launch { quizRepository.cacheQuiz(quiz) }
+        beginQuiz()
     }
 
     private fun nextQuestion() {
@@ -76,8 +104,10 @@ class QuizViewModel(menuItemType: Int) : ViewModel() {
     }
 
     fun onQuestionAnswered(answer: String) {
-        quiz.onQuestionAnswered(answer)
-        nextQuestion()
+      quiz.onQuestionAnswered(answer)
+      _answerResult.value = quiz.answerResponse()
+      //viewModelScope.launch { quizRepository.cacheQuizState(quiz) }
+      nextQuestion()
     }
 
     private fun onTimeExpired() {
@@ -87,11 +117,15 @@ class QuizViewModel(menuItemType: Int) : ViewModel() {
 
     private fun onQuizComplete() {
         val quizResult = quiz.quizComplete()
-        postQuizResult(quizResult)
+        viewModelScope.launch { quizRepository.postQuizResult(quizResult) }
+        _quizScore.value = quizResult.score
     }
 
+
     fun fragmentOnPauseCalled() {
-        quiz.pauseQuiz()
+        if (quizIsStarted) {
+            quiz.pauseQuiz()
+        }
     }
 
     fun fragmentOnResumeCalled() {
@@ -101,12 +135,16 @@ class QuizViewModel(menuItemType: Int) : ViewModel() {
     }
 
     fun fragmentOnStopCalled() {
-        //TODO cache quiz
+        //viewModelScope.launch { quizRepository.cacheQuizState(quiz) }
+    }
+
+    private fun errorRetrievingItems() {
+        _errorRetrievingData.value = true
     }
 
     private fun observeTimer() {
         viewModelScope.launch {
-            quiz.timer.timeFlow.collect { it
+            quiz.timer.timeFlow.collect {
                 _timeRemaining.value = it
                 if (it == 0) {
                     onTimeExpired()
@@ -115,105 +153,9 @@ class QuizViewModel(menuItemType: Int) : ViewModel() {
         }
     }
 
-
-
     override fun onCleared() {
         super.onCleared()
         viewModelScope.cancel()
         quiz.quizComplete()
     }
-
-    private fun getCheesecakes(): List<Cheesecake> {
-        var cheesecakeList = listOf<Cheesecake>()
-        viewModelScope.launch {
-            val response = try {
-                RetrofitInstance.airtableApi.getAllCheesecakes()
-            } catch (e: IOException) {
-                Log.e(TAG, "IOException")
-                return@launch
-            } catch (e: HttpException) {
-                Log.e(TAG, "HttpException")
-                return@launch
-            }
-
-            if (response.isSuccessful && response.body() != null) {
-                val apiCheesecakesList = response.body()
-                apiCheesecakesList?.cheesecakes?.let { cheesecakes ->
-                   cheesecakeList = ListMapperImpl(ApiCheesecakeMapper()).mapToDomain(cheesecakes)
-                    Log.i(TAG,cheesecakeList.toString())
-                    setupQuiz(cheesecakeList) //TODO fix this, this call shouldn't happen here. Figure out waiting for coroutine to finish or something
-                }
-            }
-        }
-        return cheesecakeList
-    }
-
-    private fun getDesserts(): List<Dessert> {
-        var dessertList = listOf<Dessert>()
-        viewModelScope.launch() {
-            val response = try {
-                RetrofitInstance.airtableApi.getAllDesserts()
-            } catch (e: IOException) {
-                Log.e(TAG, "IOException")
-                return@launch
-            } catch (e: HttpException) {
-                Log.e(TAG, "HttpException")
-                return@launch
-            }
-
-            if (response.isSuccessful && response.body() != null) {
-                val apiDessertList = response.body()
-                apiDessertList?.desserts?.let { desserts ->
-                    dessertList = ListMapperImpl(ApiDessertMapper()).mapToDomain(desserts)
-                    setupQuiz(dessertList) //TODO Fix this, don't call it here.
-                }
-            }
-        }
-        return dessertList
-    }
-
-    private fun getDrinks(): List<Drink> {
-        var drinkList = listOf<Drink>()
-        viewModelScope.launch {
-            val response = try {
-                RetrofitInstance.airtableApi.getAllDrinks()
-            } catch (e: IOException) {
-                Log.e(TAG,"IOException")
-                return@launch
-            } catch (e: HttpException) {
-                Log.e(TAG,"HttpException")
-                return@launch
-            }
-            if (response.isSuccessful && response.body() != null) {
-                val apiDrinkList = response.body()
-                apiDrinkList?.drinks?.let { drinks ->
-                   drinkList = ListMapperImpl(ApiDrinkMapper()).mapToDomain(drinks)
-                }
-            }
-        }
-        return drinkList
-    }
-
-    private fun postQuizResult(quizResult: QuizResult) {
-        val apiQuizResult = QuizResultToApiQuizResultMapper().mapToDomain(quizResult)
-        viewModelScope.launch {
-            val response = try {
-                RetrofitInstance.airtableApi.postQuizResult(apiQuizResult)
-            } catch (e: IOException) {
-                Log.e(TAG, "IOException")
-                return@launch
-            } catch (e: HttpException) {
-                Log.e(TAG, "HttpException")
-                return@launch
-            }
-        }
-    }
 }
-
-//TODO store all the cheesecakes, questions and answers in Room
-//TODO COMMENTS!
-//TODO reminder notifications
-//TODO share results on social media option
-//TODO if you get it wrong, allow yourself to keep trying but gray/red the incorrectly chosen answer
-
-
