+++ app/src/main/res/layout/fragment_flashcard_details_dessert.xml
@@ -95,7 +95,7 @@
         <TextView
             android:id="@+id/details_dessert_base_property_textview"
             android:layout_width="0dp"
-            android:layout_height="40dp"
+            android:layout_height="wrap_content"
             android:layout_marginBottom="16dp"
             android:textColor="@color/list_text"
             android:textSize="20sp"
@@ -104,7 +104,7 @@
             app:layout_constraintHorizontal_bias="0.5"
             app:layout_constraintStart_toStartOf="@+id/details_dessert_dishes_textview_property"
             app:layout_constraintTop_toBottomOf="@+id/details_dessert_dishes_textview_property"
-            tools:text="base property" />
+            tools:text="2 level grey scoops of apple filling" />
 
         <TextView
             android:id="@+id/details_dessert_name_textview_property"
