+++ app/src/main/res/layout/fragment_quiz_setup.xml
@@ -31,23 +31,25 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="2dp"
-        android:text="@string/chocolate"
-        android:textColor="@color/white"
-        android:textSize="18sp"
+        android:text="@string/cheesecakes"
+        android:textColor="@color/list_text"
+        android:textSize="24sp"
         app:layout_constraintBottom_toTopOf="@+id/chocolate_button"
         app:layout_constraintEnd_toEndOf="@id/chocolate_button"
         app:layout_constraintStart_toStartOf="@id/chocolate_button" />
 
     <ImageButton
         android:id="@+id/chocolate_button"
-        android:layout_width="200dp"
-        android:layout_height="150dp"
+        android:layout_width="330dp"
+        android:layout_height="250dp"
         android:layout_marginTop="42dp"
+        android:layout_marginLeft="@dimen/large_margin"
+        android:layout_marginRight="@dimen/large_margin"
         android:background="@color/velvet"
         android:contentDescription="@string/chocolate_cake"
-        android:padding="2dp"
+        android:padding="1dp"
         android:scaleType="fitCenter"
-        android:src="@drawable/herysheys_cheesecake"
+        android:src="@drawable/salted_caramel"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/select_category_text" />
@@ -57,23 +59,23 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_marginBottom="2dp"
-        android:text="@string/fruit"
-        android:textColor="@color/white"
-        android:textSize="18sp"
+        android:text="@string/desserts"
+        android:textColor="@color/list_text"
+        android:textSize="24sp"
         app:layout_constraintBottom_toTopOf="@+id/fruit_button"
         app:layout_constraintEnd_toEndOf="@id/fruit_button"
         app:layout_constraintStart_toStartOf="@id/fruit_button" />
 
     <ImageButton
         android:id="@+id/fruit_button"
-        android:layout_width="200dp"
-        android:layout_height="150dp"
+        android:layout_width="330dp"
+        android:layout_height="250dp"
         android:layout_marginTop="36dp"
         android:background="@color/velvet"
         android:contentDescription="@string/chocolate_cake"
-        android:padding="2dp"
+        android:padding="1dp"
         android:scaleType="fitCenter"
-        android:src="@drawable/lemon_raspberry_cheesecake"
+        android:src="@drawable/godivabrowniesundae"
         app:layout_constraintEnd_toEndOf="@+id/chocolate_button"
         app:layout_constraintStart_toStartOf="@+id/chocolate_button"
         app:layout_constraintTop_toBottomOf="@+id/chocolate_button" />
@@ -84,8 +86,9 @@
         android:layout_height="wrap_content"
         android:layout_marginBottom="2dp"
         android:text="@string/other"
-        android:textColor="@color/white"
+        android:textColor="@color/list_text"
         android:textSize="18sp"
+        android:visibility="gone"
         app:layout_constraintBottom_toTopOf="@+id/other_button"
         app:layout_constraintEnd_toEndOf="@id/other_button"
         app:layout_constraintStart_toStartOf="@id/other_button" />
@@ -100,6 +103,7 @@
         android:padding="2dp"
         android:scaleType="fitCenter"
         android:src="@drawable/tiramisu_cheesecake"
+        android:visibility="gone"
         app:layout_constraintEnd_toEndOf="@+id/fruit_button"
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="@+id/fruit_button"
