+++ app/src/main/res/layout/list_item.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="wrap_content"
+    android:background="@color/list_background">
+
+    <ImageView
+        android:id="@+id/list_item_imageview"
+        android:layout_width="422dp"
+        android:layout_height="289dp"
+        android:paddingTop="@dimen/default_margin"
+        app:layout_constraintTop_toBottomOf="@id/list_item_title_textview"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:srcCompat="@drawable/fresh_strawberry_cheesecake" />
+
+    <TextView
+        android:id="@+id/list_item_title_textview"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingStart="@dimen/large_margin"
+        android:paddingTop="@dimen/small_margin"
+        android:paddingBottom="@dimen/small_margin"
+        android:textColor="@color/list_text"
+        android:textSize="@dimen/large_string_size"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:text="Item Name" />
+</androidx.constraintlayout.widget.ConstraintLayout>
