+++ app/src/main/res/navigation/nav_graph.xml
@@ -50,6 +50,10 @@
         <action
             android:id="@+id/action_quizFragment_to_quizResultsFragment"
             app:destination="@id/quizResultsFragment" />
+        <action
+            android:id="@+id/action_quizFragment_to_quizResultsSimpleFragment"
+            app:destination="@id/quizResultsSimpleFragment"
+            app:popUpTo="@id/mainMenuFragment" />
     </fragment>
     <fragment
         android:id="@+id/quizResultsFragment"
@@ -100,6 +104,18 @@
         tools:layout="@layout/fragment_flashcard_details" >
         <argument
             android:name="menuItemType"
+            app:argType="integer" />
+    </fragment>
+    <fragment
+        android:id="@+id/quizResultsSimpleFragment"
+        android:name="com.example.mycheesecakes.ui.quiz.QuizResultsSimpleFragment"
+        android:label="QuizResultsSimpleFragment"
+        tools:layout="@layout/fragment_quiz_result_simple">
+        <argument
+            android:name="numCorrect"
+            app:argType="integer" />
+        <argument
+            android:name="total"
             app:argType="integer" />
     </fragment>
 </navigation>
